Reorganized structure to have separate test folders. Added initial maxdistancewrapper.
