+++ pom.xml
@@ -61,6 +61,11 @@
             <artifactId>scalatest_2.10</artifactId>
             <version>2.0.M5b</version>
         </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.0</version>
+        </dependency>
         <!-- other biojava jars as needed -->
         <dependency>
             <groupId>net.sourceforge.argparse4j</groupId>
