+++ src/main/edu/gsu/cs/kgem/model/EM.scala
