+++ src/main/edu/gsu/cs/kgem/model/Genotype.scala
